requirements.txt
langchain
ipykernel
python-dotenv
arxiv
bs4
fitz
pymupdf
langchain-text-splitters
langchain-community
langchain-ollama
langchain-huggingface
sentence_transformers
faiss-cpu
langchain_chroma
python-dotenv
streamlit
langserve
langchain-groq
uvicorn[standard]
fastapi[standard]
sse_starlette
selenium
unstructured
validators
pdfminer.six
pi_heif
unstructured-inference
pdf2image
langgraph
mysql.connector
mysql-connector-python
wikipedia
numexpr
huggingface_hub
cassio
tiktoken
datasets
PyPDF2
PdfReader
gradio